restructured project into "mbi" and "header" modules
