--- src/tags/mod.rs
@@ -1,4 +1,0 @@
-//! Multiboot2 knows about two different kinds of tags.
-
-pub mod header;
-pub mod mbi;
